+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * `Utc2k::cmp_time`
 * `Utc2k::from_smooshed_date_str`
 * `Utc2k::from_smooshed_datetime_str`
+* `Utc2k::to_midnight`